backend/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
datasets==3.3.2
dill==0.3.8
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
faiss-cpu==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.19
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mysql-connector-python==9.2.0
mysqlclient==2.2.7
networkx==3.4.2
nltk==3.9.1
numpy>=1.26,<2.0
openai==1.64.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
